parsing_ppm sep file
