+++ img/cat_ok_delim.ppm
@@ -18,7 +18,7 @@ P3
  
 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117
  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26
- 69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69
+ 69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69# this is comment!!
 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117
  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26
  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69 117  26  69
